renames peer_observer mod
